Extract IntervalCard content as separate CardContent component
